# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SahaayX — Smart Resource Allocation & Volunteer Coordination Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /volunteers:
    get:
      operationId: listVolunteers
      tags: [volunteers]
      summary: List volunteers
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: skill
          in: query
          schema:
            type: string
        - name: district
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of volunteers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Volunteer"
    post:
      operationId: createVolunteer
      tags: [volunteers]
      summary: Register a volunteer profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVolunteerBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Volunteer"
  /volunteers/{id}:
    get:
      operationId: getVolunteer
      tags: [volunteers]
      summary: Get volunteer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Volunteer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Volunteer"
        "404":
          description: Not found
    put:
      operationId: updateVolunteer
      tags: [volunteers]
      summary: Update volunteer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVolunteerBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Volunteer"
  /volunteers/{id}/qr:
    get:
      operationId: getVolunteerQr
      tags: [volunteers]
      summary: Get volunteer QR code data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: QR data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VolunteerQrData"
...
  /resources:
    get:
      operationId: listResources
      tags: [resources]
      summary: List resources
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: district
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Resources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resource"
    post:
      operationId: createResource
      tags: [resources]
      summary: Create a resource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResourceBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
  /resources/{id}:
    put:
      operationId: updateResource
      tags: [resources]
      summary: Update resource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateResourceBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List impact reports
      responses:
        "200":
          description: Reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ImpactReport"
    post:
      operationId: createReport
      tags: [reports]
      summary: Submit an incident/impact report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReportBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImpactReport"
  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get report by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImpactReport"
...
components:
  schemas:
...
    Volunteer:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        district:
          type: string
        skills:
          type: array
          items:
            type: string
        availability:
          type: string
        status:
          type: string
          enum: [active, inactive, deployed]
        totalHours:
          type: number
        missionsCompleted:
          type: integer
        badge:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - district
        - skills
        - status
        - totalHours
        - missionsCompleted
        - createdAt
    CreateVolunteerBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        district:
          type: string
        skills:
          type: array
          items:
            type: string
        availability:
          type: string
      required:
        - name
        - email
        - district
        - skills
    UpdateVolunteerBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        district:
          type: string
        skills:
          type: array
          items:
            type: string
        availability:
          type: string
        status:
          type: string
    VolunteerQrData:
      type: object
      properties:
        volunteerId:
          type: integer
        name:
          type: string
        district:
          type: string
        skills:
          type: array
          items:
            type: string
        badge:
          type: string
        qrPayload:
          type: string
      required:
        - volunteerId
        - name
        - district
        - skills
        - qrPayload
...
    Resource:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [medical, food, shelter, transport, equipment, manpower]
        district:
          type: string
        quantity:
          type: integer
        unit:
          type: string
        status:
          type: string
          enum: [available, allocated, depleted]
        allocatedToMissionId:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - district
        - quantity
        - unit
        - status
        - createdAt
    CreateResourceBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        district:
          type: string
        quantity:
          type: integer
        unit:
          type: string
      required:
        - name
        - category
        - district
        - quantity
        - unit
    UpdateResourceBody:
      type: object
      properties:
        quantity:
          type: integer
        status:
          type: string
        allocatedToMissionId:
          type: integer
    ImpactReport:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [incident, impact, needs_assessment, compliance]
        district:
          type: string
        content:
          type: string
        submittedBy:
          type: string
        status:
          type: string
          enum: [pending, reviewed, approved, rejected]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - type
        - district
        - content
        - submittedBy
        - status
        - createdAt
    CreateReportBody:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
          enum: [incident, impact, needs_assessment, compliance]
        district:
          type: string
        content:
          type: string
      required:
        - title
        - type
        - district
        - content
...