# lib/api-zod/src/generated/types/createResourceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SahaayX — Smart Resource Allocation & Volunteer Coordination Platform
 * OpenAPI spec version: 0.1.0
 */

export interface CreateResourceBody {
  name: string;
  category: string;
  district: string;
  quantity: number;
  unit: string;
}