# lib/api-zod/src/generated/types/impactReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SahaayX — Smart Resource Allocation & Volunteer Coordination Platform
 * OpenAPI spec version: 0.1.0
 */
import type { ImpactReportStatus } from "./impactReportStatus";
import type { ImpactReportType } from "./impactReportType";

export interface ImpactReport {
  id: number;
  title: string;
  type: ImpactReportType;
  district: string;
  content: string;
  submittedBy: string;
  status: ImpactReportStatus;
  createdAt: Date;
}