# lib/api-zod/src/generated/types/resource.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SahaayX — Smart Resource Allocation & Volunteer Coordination Platform
 * OpenAPI spec version: 0.1.0
 */
import type { ResourceCategory } from "./resourceCategory";
import type { ResourceStatus } from "./resourceStatus";

export interface Resource {
  id: number;
  name: string;
  category: ResourceCategory;
  district: string;
  quantity: number;
  unit: string;
  status: ResourceStatus;
  allocatedToMissionId?: number;
  createdAt: Date;
}